chore(member): reorganized blog hook folder structure (#152)
